apps/sector_classification_model_train/requirements.txt
aws-cdk-lib==2.149.0
aws-cdk.aws-lambda-python-alpha
constructs>=10.0.0,<11.0.0
python-dotenv
